Move MessageTemplate to a separate file
