+++ mailmerge/__init__.py
@@ -3,6 +3,3 @@ Mail merge module importable API.
 
 Andrew DeOrio <awdeorio@umich.edu>
 """
-
-from .api import sendall
-from .api import MessageTemplate
